feat(status-filter): moved state management to state
