+++ src/components/common/loads-wrapper/filters-dropdown/index.tsx
@@ -1,21 +1,16 @@
-import { useState } from "react"
-
 import { Button } from "@/components/ui/button"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { useLoadsContext } from "@/contexts/loads-context"
 
 export const FiltersDropdown = () => {
-  const { setSearchQuery } = useLoadsContext()
-  const [statusFilter, setStatusFilter] = useState<string>("")
+  const { statusFilter, setStatusFilter } = useLoadsContext()
 
   const handleFilterChange = (value: string) => {
     setStatusFilter(value)
-    setSearchQuery(value ? `status:${value}` : "")
   }
 
   const resetFilter = () => {
     setStatusFilter("")
-    setSearchQuery("")
   }
 
   return (
@@ -31,9 +26,9 @@ export const FiltersDropdown = () => {
         </SelectContent>
       </Select>
       {statusFilter && (
-        <Button 
-          variant="ghost" 
-          size="sm" 
+        <Button
+          variant="ghost"
+          size="sm"
           onClick={resetFilter}
           className="h-8 px-2 text-sm text-slate-500 hover:text-slate-900"
         >
