+++ src/utils/loads-context/filter-loads/index.ts
@@ -1,21 +1,22 @@
 import { Load } from "@/types/load"
 
-export const filterLoads = (loads: Load[], searchQuery: string): Load[] => {
-  if (!searchQuery) return loads
+export const filterLoads = (loads: Load[], searchQuery: string, statusFilter: string): Load[] => {
+  let filteredLoads = loads
 
-  // Check if the search query is a status filter
-  if (searchQuery.startsWith('status:')) {
-    const status = searchQuery.replace('status:', '').toLowerCase()
-    return loads.filter(load => load.status.toLowerCase() === status)
+  if (statusFilter) {
+    filteredLoads = filteredLoads.filter(load => load.status.toLowerCase() === statusFilter.toLowerCase())
   }
 
-  // Regular search across all fields
-  return loads.filter(load => 
-    load.id.toString().includes(searchQuery) ||
-    load.status.toLowerCase().includes(searchQuery.toLowerCase()) ||
-    load.origin.toLowerCase().includes(searchQuery.toLowerCase()) ||
-    load.destination.toLowerCase().includes(searchQuery.toLowerCase()) ||
-    load.client_name.toLowerCase().includes(searchQuery.toLowerCase()) ||
-    load.carrier_name.toLowerCase().includes(searchQuery.toLowerCase())
-  )
+  if (searchQuery) {
+    filteredLoads = filteredLoads.filter(load => 
+      load.id.toString().includes(searchQuery) ||
+      load.status.toLowerCase().includes(searchQuery.toLowerCase()) ||
+      load.origin.toLowerCase().includes(searchQuery.toLowerCase()) ||
+      load.destination.toLowerCase().includes(searchQuery.toLowerCase()) ||
+      load.client_name.toLowerCase().includes(searchQuery.toLowerCase()) ||
+      load.carrier_name.toLowerCase().includes(searchQuery.toLowerCase())
+    )
+  }
+
+  return filteredLoads
 }
